+++ DataLayer/BookApp/EfCode/Configurations/BookConfig.cs
@@ -12,43 +12,33 @@ namespace DataLayer.BookApp.EfCode.Configurations
         public void Configure
             (EntityTypeBuilder<Book> entity)
         {
-            entity.Property(p => p.PublishedOn)//#A
+            entity.Property(p => p.PublishedOn)
                 .HasColumnType("date");        
 
-            entity.Property(p => p.Price) //#B
+            entity.Property(p => p.Price) 
                 .HasColumnType("decimal(9,2)");
 
-            entity.Property(x => x.ImageUrl) //#C
+            entity.Property(x => x.ImageUrl) 
                 .IsUnicode(false);
 
-            entity.HasIndex(x => x.PublishedOn); //#D
+            entity.HasIndex(x => x.PublishedOn); 
 
             //Model-level query filter
 
             entity
-                .HasQueryFilter(p => !p.SoftDeleted); //#E
+                .HasQueryFilter(p => !p.SoftDeleted); 
 
             //----------------------------
             //relationships
 
-            entity.HasOne(p => p.Promotion) //#A
-                .WithOne() //#A
-                .HasForeignKey<PriceOffer>(p => p.BookId); //#A
+            entity.HasOne(p => p.Promotion) 
+                .WithOne() 
+                .HasForeignKey<PriceOffer>(p => p.BookId); 
 
-            entity.HasMany(p => p.Reviews)     //#B
-                .WithOne()                     //#B
-                .HasForeignKey(p => p.BookId); //#B
+            entity.HasMany(p => p.Reviews)     
+                .WithOne()                     
+                .HasForeignKey(p => p.BookId); 
         }
     }
-    /*Type/Size setting**********************************************
-    #A The convention-based mapping for .NET DateTime is SQL datetime2. This command changes the SQL column type to date, which only holds the date, not time
-    #B I set a smaller precision and scale of (9,2) for the price instead of the default (18,2)
-    #C The convention-based mapping for .NET string is SQL nvarchar (16 bit Unicode). This command changes the SQL column type to varchar (8 bit ASCII)
-    #D I add an index to the PublishedOn property because I sort and filter on this property
-    #E This sets a model-level query filter on the Book entity. By default, a query will exclude Book entites where th SoftDeleted property is true
-     * * ******************************************************/
-    /*CH07********************************************************
-    #A This defines the One-to-One relationship to the promotion that a book can optionally have. The foreign key is in the PriceOffer
-    #B This defines the One-to-Many relationship, with a book having zero to many reviews
-     * ***********************************************************/
+
 }